+++ packages/ve-table/src/footer/footer-td.jsx
@@ -303,6 +303,10 @@ export default {
         cellMousemove(e, fn) {
             fn && fn(e);
         },
+        // mouseover
+        cellMouseover(e, fn) {
+            fn && fn(e);
+        },
         // mousedown
         cellMousedown(e, fn) {
             fn && fn(e);
@@ -336,6 +340,7 @@ export default {
             mouseenter,
             mouseleave,
             mousemove,
+            mouseover,
             mousedown,
             mouseup,
         } = customEvents;
@@ -358,6 +363,9 @@ export default {
             },
             mousemove: (e) => {
                 this.cellMousemove(e, mousemove);
+            },
+            mouseover: (e) => {
+                this.cellMouseover(e, mouseover);
             },
             mousedown: (e) => {
                 this.cellMousedown(e, mousedown);